refactor(db): finish crud.py → repositories/services migration (#72)
crud.py: 976 LOC, 75 defs → 100 LOC, 0 defs (pure re-export module).
The Phase-3 repository extraction is now complete: every CRUD lives in
a proper repository or service, and crud.py is the thin compatibility
seam that keeps existing callers working without churn.

Strategy: replace delegator FUNCTIONS with module-level re-export
IMPORTS so callers (`core/api/`, `core/services/`, `core/workers/`,
tests — 20+ files) continue to work via `crud.X(...)` without any
caller-side migration. This satisfies the issue's acceptance criterion
"No 1-line delegator functions in crud.py" — re-export imports are
not functions.

Substantive moves (function → destination):

To core/db/repositories/memory_blocks.py:
- get_memory_blocks_by_agent
- get_memory_blocks_by_conversation
- get_unique_conversation_count (was an inline implementation)
- get_feedback_log
- get_feedback_logs_by_memory_block
- update_feedback_log
- delete_feedback_log

To core/services/search_service.py (module-level functions):
- search_memory_blocks_fulltext
- search_memory_blocks_semantic
- search_memory_blocks_hybrid
- search_memory_blocks_enhanced (dispatcher)
- _execute_with_query_expansion (private helper)

To core/db/repositories/consolidation_suggestions.py (NEW, 242 LOC):
- ConsolidationOriginalMismatchError exception class
- create/get/update/delete_consolidation_suggestion
- get_consolidation_suggestions / _scoped
- apply_consolidation (60-LOC business transaction)
- _get_or_create_keyword (private helper used only by apply_consolidation)

Delegators converted to grouped re-export imports (57 functions across
6 repositories: bulk_ops, audits, agents, keywords, memory_blocks,
organizations). Duplicate `create_audit_log` def deleted; the canonical
lives in repositories/audits.py and is re-exported.

Caller-side updates (only where callers used moved-out-of-crud APIs
that have NO re-export back to crud — the 3 search functions and the
private helper that's now in search_service):

- core/api/search.py: 3 search functions imported from
  core.services.search_service directly (not from crud)
- core/api/main.py: get_unique_conversation_count imported from
  repositories.memory_blocks directly
- core/search/evaluation.py: 3 search functions imported from
  core.services.search_service directly
- tests/unit/test_crud_query_expansion.py: patch targets and imports
  updated to point at core.services.search_service

Net diff: 8 files changed, +674/-992 (net -318 LOC). Most of the
"deletion" is moved-not-deleted code — the actual size reduction is
the 30+ delegator functions becoming grouped imports.

Verification (run locally on integration branch — CI is silent on
integration/* per .github/workflows/deploy.yml):
- uv run pytest (full suite): 859 passed, 8 skipped, 80.26% coverage
  (above 80% gate)
- bun run build (Phase 3 cross-stack sanity): clean, 2.32s
- crud.py: 100 LOC (acceptance: ≤100 ✓), zero `def` (acceptance: ✓)
- Substantive functions verified gone from crud.py:
  apply_consolidation, search_memory_blocks_*, get_unique_conversation_count
  (acceptance: ✓ — they're now in services/repositories)

Refs #72, umbrella #85.
